+++ tests/vppinputoutput.h
@@ -80,7 +80,7 @@ public:
 class PooledFrameAllocator : public FrameAllocator
 {
 public:
-    PooledFrameAllocator(const SharedPtr<VADisplay>& display, int poolsize):
+    PooledFrameAllocator(const SharedPtr<VADisplay>& display, size_t poolsize):
         m_display(display), m_poolsize(poolsize)
     {
     }
@@ -108,7 +108,7 @@ public:
             return false;
         }
         std::deque<SharedPtr<VideoFrame> > buffers;
-        for (int i = 0;  i < m_surfaces.size(); i++) {
+        for (size_t i = 0;  i < m_surfaces.size(); i++) {
             SharedPtr<VideoFrame> f(new VideoFrame);
             memset(f.get(), 0, sizeof(VideoFrame));
             //we need fill dest crop to work around libva's bug.
@@ -133,7 +133,7 @@ private:
     SharedPtr<VADisplay> m_display;
     std::vector<VASurfaceID> m_surfaces;
     SharedPtr<VideoPool<VideoFrame> > m_pool;
-    int m_poolsize;
+    size_t m_poolsize;
 };
 
 class VaapiFrameIO
@@ -172,10 +172,10 @@ public:
             return false;
         }
         bool ret = true;
-        for (int i = 0; i < planes; i++) {
+        for (uint32_t i = 0; i < planes; i++) {
             char* ptr = buf + image.offsets[i];
             int w = byteWidth[i];
-            for (int j = 0; j < byteHeight[i]; j++) {
+            for (uint32_t j = 0; j < byteHeight[i]; j++) {
                 ret = m_io(ptr, w, fp);
                 if (!ret)
                     goto out;
@@ -209,7 +209,7 @@ private:
     SharedPtr<VaapiFrameIO> m_frameio;
     static bool readFromFile(char* ptr, int size, FILE* fp)
     {
-        return fread(ptr, 1, size, fp) == size;
+        return fread(ptr, 1, size, fp) == (size_t)size;
     }
 };
 
@@ -228,7 +228,7 @@ private:
     SharedPtr<VaapiFrameIO> m_frameio;
     static bool writeToFile(char* ptr, int size, FILE* fp)
     {
-        return fwrite(ptr, 1, size, fp) == size;
+        return fwrite(ptr, 1, size, fp) == (size_t)size;
     }
 };
 //vaapi related operation end
